pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-app"
version = "0.1.0"
description = "Codex dâhilinde kullanılmak üzere Python uygulama iskeleti"
authors = [{name="Your Name", email="your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "python-dotenv>=1.0.1",
    "openai>=1.50.0",
    "typer>=0.12.5",
]

[tool.black]
line-length = 100
target-version = ["py310","py311","py312"]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203","W503"]
